requirements.txt
langchain>=0.0.222
python_dotenv>=1.0.0,<2.0.0
yt_dlp==2023.7.6
pydub>=0.25.1,<0.26.0
openai>=0.27.8,<0.28.0